parser_python/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
fake-useragent==1.5.1
frozenlist==1.5.0
idna==3.10
lxml==5.3.0
multidict==6.1.0
propcache==0.2.0
psycopg2-binary==2.9.10
requests==2.32.3
schedule==1.2.2
soupsieve==2.6
urllib3==2.2.3
yarl==1.17.1
